% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicroBreed.R
\name{MicroBreed}
\alias{MicroBreed}
\title{Create MicroBreed Object}
\usage{
MicroBreed(genotype, phenotype, microbiome, metadata = NULL)
}
\arguments{
\item{genotype}{Data frame of genotype information}

\item{phenotype}{Data frame of phenotype information}

\item{microbiome}{Data frame of microbiome abundances}

\item{metadata}{Optional metadata}
}
\value{
A MicroBreed object
}
\description{
Store genotype, phenotype, microbiome, and metadata
}
