% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_microbiome.R
\name{read_microbiome}
\alias{read_microbiome}
\title{Read microbiome abundance table}
\usage{
read_microbiome(file)
}
\arguments{
\item{file}{CSV file}
}
\value{
data frame
}
\description{
Import microbiome abundance data.
}
